audio_receiver/requirements.txt
fastapi
uvicorn
python-dotenv
yandex-music